Fixed armature rotation being hijacked by another object with the same name in the hierarchy
If there's an object with the same name as the avatar's armature in the hierarchy somewhere, when denormalized avatar builds a (unity) humanoid avatar, it uses that armature object's rotation instead of the avatar armature's rotation.

This commonly happens when people attach skinned meshes directly to bones (like a pair of glasses that has an armature to the head bone). If the object has a weird rotation value, maybe because it was exporter with a different coordinate system, it will break the avatar's pose.

Steps to reproduce:
- On an avatar with the armature named `Armature`, create a GameObject named `Armature`.
- Parent it somewhere, like Head
- Rotate the object 90 degrees, for example
- Export
- The avatar will be rotated 90 degrees.

Tested in Warudo only.

Only fix I found is to make sure we don't have duplicate names for the hips parent in the hierarchy.